version 0.0.11+7

+++ app/Http/Controllers/OnlineNewsletterController.php
@@ -10,16 +10,25 @@ use Illuminate\View\View;
 
 class OnlineNewsletterController extends Controller
 {
+    private function message(): string {
+        $str = array(
+            'fr' => array(),
+            'en' => array(""),
+        );
+
+        return "";
+    }
     public function store(Request $request): RedirectResponse
     {
-        if(!$request->has('email')) return Redirect::back();
+        if (!$request->has('email')) return Redirect::back();
         $email = $request->input('email');
 
-        if(OnlineNewsletter::whereEmail($email)->exists())
+        if (OnlineNewsletter::whereEmail($email)->exists())
             return Redirect::back()->with([
-                'newsletter'=>'already',
-                'email'=>"$email",
-                'q' => "zHX854D5SDs8tx5#00HMsk7"
+                'newsletter' => 'already',
+                'email' => "$email",
+                'q' => "newsletter",
+                'qr' => "zHX854D5SDs8tx5#00HMsk7",
             ]);
 
         $newsletter = new OnlineNewsletter();
@@ -29,27 +38,32 @@ class OnlineNewsletterController extends Controller
         $request->session()->put('newsletter', true);
 
         return Redirect::route('home.index', [
-            'newsletter'=>'done',
-            'q'=>"saved",
+            'newsletter' => 'done',
+            'q' => "newsletter",
+            'message'=>"done",
         ]);
 
     }
 
-    public function delete(Request $request) : View | RedirectResponse
+    public function delete(Request $request): View|RedirectResponse
     {
+        /// fr/home/newsletter/unsubscribe
 
-        return view("newsletter");
+        //return view("newsletter");
 
-        /*if(!$request->has('unsubscribe') and !$request->has('email'))
-             abort(404); //Redirect::route('home.index',['unsubscribe'=>"null"]);
+        if (!$request->has('email')) abort(404);
+        //Redirect::route('home.index',['unsubscribe'=>"null"]);
 
         $email = $request->input('email');
-        if(!OnlineNewsletter::whereEmail($email)->exists())
-            return Redirect::route('home.index', ['q'=>'none']);
+        if (!OnlineNewsletter::whereEmail($email)->exists())
+            return Redirect::route('home.index', [
+                'q' => 'newsletter',
+                'message' => "email do not exist",
+            ]);
 
         OnlineNewsletter::whereEmail($email)->delete();
 
         $email = $request->input('email');
-        return view("newsletter.unsubscribe",['email'=>$email]);*/
+        return view("newsletter", ['email' => $email]);
     }
 }
